Move remove() and removeAll() to tests
This is done to keep deadcode quiet.
